// web/src/types/generated/types.gen.ts
export type FileFull = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Title
     *
     * File title
     */
    title: string;
    /**
     * Content Type
     *
     * File content type
     */
    content_type: string;
    /**
     * S3 Link
     *
     * Public S3 link
     */
    s3_link: string;
};
// ...
export type StudentInput = {
    /**
     * Fullname
     *
     * Full name
     */
    fullname: string;
    /**
     * Github
     *
     * GitHub profile
     */
    github?: string | null;
    /**
     * Telegram
     *
     * Telegram handle
     */
    telegram?: string | null;
    /**
     * Phone Number
     *
     * Phone number
     */
    phone_number?: string | null;
    /**
     * Primary Course Uid
     *
     * Primary course UID
     */
    primary_course_uid: string;
    /**
     * Primary Direction Uid
     *
     * Primary direction UID
     */
    primary_direction_uid: string;
    /**
     * Avatar Uid
     *
     * Student avatar UID
     */
    avatar_uid?: string | null;
    /**
     * Education Start Date
     *
     * Date when educations started (or should be started)
     */
    education_start_date: string;
    /**
     * Education Finish Date
     *
     * Date when education was or will be finished
     */
    education_finish_date: string;
    /**
     * Birthday Date
     *
     * Birthday date
     */
    birthday_date: string;
    /**
     * Is Finished
     *
     * Education finished flag
     */
    is_finished: boolean;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived?: boolean;
    /**
     * User Uid
     *
     * Linked user UID
     */
    user_uid?: string | null;
};

/**
 * student.StudentListedOutput
 */
export type StudentListedOutput = {
    /**
     * Items
     *
     * List of students
     */
    items: Array<StudentOutput>;
    /**
     * Count
     *
     * Total count
     */
    count: number;
};

/**
 * student.StudentOutput
 */
export type StudentOutput = {
    /**
     * Uid
     *
     * UID
     */
    uid: string;
    /**
     * Fullname
     *
     * Full name
     */
    fullname: string;
    /**
     * Github
     *
     * GitHub profile
     */
    github?: string | null;
    /**
     * Telegram
     *
     * Telegram handle
     */
    telegram?: string | null;
    /**
     * Phone Number
     *
     * Phone number
     */
    phone_number?: string | null;
    /**
     * Primary Course Uid
     *
     * Primary course UID
     */
    primary_course_uid: string;
    /**
     * Primary Direction Uid
     *
     * Primary direction UID
     */
    primary_direction_uid: string;
    /**
     * Avatar Uid
     *
     * Student avatar UID
     */
    avatar_uid?: string | null;
    /**
     * Education Start Date
     *
     * Date when educations started (or should be started)
     */
    education_start_date: string;
    /**
     * Education Finish Date
     *
     * Date when education was or will be finished
     */
    education_finish_date: string;
    /**
     * Birthday Date
     *
     * Birthday date
     */
    birthday_date: string;
    /**
     * Created Date
     *
     * Created date
     */
    created_date: string;
    /**
     * Is Finished
     *
     * Education finished flag
     */
    is_finished: boolean;
    /**
     * Is Archived
     *
     * Archive flag
     */
    is_archived: boolean;
    /**
     * User Uid
     *
     * Linked user UID
     */
    user_uid?: string | null;
    /**
     * Student avatar
     */
    avatar?: FileFull | null;
};